// src/app/core/services/model/controllerPointOutputDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Controller Point data transfer object for frontend representation
 */
export interface ControllerPointOutputDto { 
    /**
     * UUID of device point
     */
    uuid?: string;
    /**
     * device point name
     */
    name?: string;
    /**
     * Controller ID of a device point
     */
    controllerId?: number;
    /**
     * remark of device point
     */
    remark?: string;
    /**
     * Enabled timestamp of a device point
     */
    enabledTimestamp?: Date;
    /**
     * Disabled timestamp of a device point
     */
    disabledTimestamp?: Date;
    /**
     * Clique Name
     */
    cliqueName?: string;
    /**
     * Controller UUID
     */
    controller_UUID?: string;
    /**
     * Point Address UUID
     */
    pointAddress_UUID?: string;
    /**
     * Managed By User UUID
     */
    managedByUser_UUID?: string;
    /**
     * uuid != GlobalDataMgr.CurrentControllerPointUuid
     */
    isRemoteController?: boolean;
}