// src/app/core/services/model/devicePointCreateInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * data transfer object for creating a new device point in frontend representation
 */
export interface DevicePointCreateInput { 
    /**
     * Name of a Device Point
     */
    name: string;
    /**
     * IP-Address for a hyplus device point
     */
    ipAddress?: string;
    /**
     * SerialNumber for a hyplus device
     */
    serialNumber?: number;
    /**
     * RevisionNumber for a hyplus device
     */
    revisionNumber?: number;
    /**
     * Remark of Device Point
     */
    remark?: string;
    /**
     * UUID of PointAddress
     */
    pointAddressUUID: string;
    /**
     * Name of the device type
     */
    deviceTypeName: string;
    /**
     * reload Config (defaut: true)
     */
    reloadConfig?: boolean;
}